pyproject.toml
[project]
name = "robotics-task-horizon"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.96.0",
    "matplotlib>=3.10.8",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "tqdm>=4.67.3",
]
